+++ BackEnd/repository/UserLabel.js
@@ -1,7 +1,8 @@
 const db = require('../database/models')
+const { OTJSDBEntityNotFoundException } = require('../Exceptions/OTJSErrors')
 
 class UserLabel{
-  static findOne(user_id,label_name){
+  static getUserLabel(user_id,label_name){
     return db.UserLabel.findOne({
       where:{
         user_id:user_id,
@@ -10,14 +11,19 @@ class UserLabel{
     })
   }
 
-  static create(user_id,_label_name){
+  static create(user_id,label_name){
     return db.UserLabel.create({
       user_id:user_id,
       label_name:label_name
     })
   }
 
-  static destroy(user_id,label_name){
+  static delete(user_id,label_name){
+    const user_label = UserLabel.getUserLabel(user_id,label_name)
+    if(user_label==null){
+      throw new OTJSDBEntityNotFoundException('This UserLabel doesn\'t exist')
+    }
+
     return db.UserLabel.destroy({
       where:{
         user_id:user_id,
@@ -26,7 +32,7 @@ class UserLabel{
     })
   }
 
-  static findAll(){
+  static getAllUserLabel(){
     return db.UserLabel.findAll()
   }
 }
